+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-11-08 02:55:18
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: aa523fe335a895874aa12250514799371e292366
+
+## [ForwardWidgets-2kuai] - 2025-11-08 02:55:18
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 3feeb7d989c9f43c4d6a687da00daaa7a80c6fb1
+
+## [ForwardWidgets-pack1r] - 2025-11-08 02:55:18
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: f462e6bce322a39904e52da75659dff0bcdc7fe6
+
+
 ## [ForwardWidgets-huangxd] - 2025-11-07 03:03:06
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main